improved performance and accuracy - changed life calc and added stopwatch

+++ TownConquer/Server/Game_Server/EA/Models/Individual.cs
@@ -13,11 +13,11 @@ namespace Game_Server.EA.Models {
         public Vector3 startPos;
         public bool isElite { get; set; }
         public int score = 0;
-        public List<int> timestamp;
+        public List<long> timestamp;
 
         public Individual(int number) {
             this.number = number;
-            timestamp = new List<int>();
+            timestamp = new List<long>();
         }
 
         /// <summary>
